+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 
 - Stats API. (`#72`_)
 - Sender Signatures API. (`#73`_)
+- Messages API. (`#71`_)
 
 0.6.2 - 02.01.2017
 ------------------
@@ -99,6 +100,7 @@ Changelog
 .. _#76: https://github.com/Stranger6667/postmarker/issues/76
 .. _#73: https://github.com/Stranger6667/postmarker/issues/73
 .. _#72: https://github.com/Stranger6667/postmarker/issues/72
+.. _#71: https://github.com/Stranger6667/postmarker/issues/71
 .. _#64: https://github.com/Stranger6667/postmarker/issues/64
 .. _#62: https://github.com/Stranger6667/postmarker/issues/62
 .. _#60: https://github.com/Stranger6667/postmarker/issues/60